# api/docs/swagger.yaml
definitions:
  handler.Response:
    properties:
      data: {}
      description:
        type: string
      error: {}
      status:
        type: integer
    type: object
  models.Contact:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      first_name:
        type: string
      id:
        type: string
      last_name:
        type: string
      middle_name:
        type: string
      phone_number:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  models.CreateContact:
    properties:
      first_name:
        description: UserID      string `json:"user_id"`
        type: string
      last_name:
        type: string
      middle_name:
        type: string
      phone_number:
        type: string
    type: object
...
  models.UpdateContact:
    properties:
      first_name:
        description: UserID      string `json:"user_id"`
        type: string
      last_name:
        type: string
      middle_name:
        type: string
      phone_number:
        type: string
    type: object
...
info:
  contact: {}
  description: This is a sample server.
  title: Swagger Example API
  version: "1.0"
paths:
  /task/api/v1/createcontact:
    post:
      consumes:
      - application/json
      description: Create a new contact
      operationId: create_contact
      parameters:
      - description: Contact
        in: body
        name: Contact
        required: true
        schema:
          $ref: '#/definitions/models.CreateContact'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully created contact
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Create Contact
      tags:
      - contact
...
  /task/api/v1/deletecontact/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a contact by its ID
      operationId: delete_Contact
      parameters:
      - description: Contact ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully deleted
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Delete contact by ID
      tags:
      - contact
...
  /task/api/v1/getallcontacts:
    get:
      consumes:
      - application/json
      description: Retrieve all Contacts
      operationId: get_all_Contacts
      parameters:
      - description: Search Contacts by name or email
        in: query
        name: search
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Limit number of results per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved contacts
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "404":
          description: Contact not found
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Get all Contacts
      tags:
      - contact
...
  /task/api/v1/getbyidcontact/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve a contact by their ID
      operationId: get_Contact
      parameters:
      - description: Contact ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved contact
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "404":
          description: Contact not found
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Get contact by ID
      tags:
      - contact
...
  /task/api/v1/updatecontact/{id}:
    put:
      consumes:
      - application/json
      description: Update an existing contact
      operationId: update_contact
      parameters:
      - description: Contact ID
        in: path
        name: id
        required: true
        type: string
      - description: UpdateContactRequest
        in: body
        name: Contact
        required: true
        schema:
          $ref: '#/definitions/models.UpdateContact'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Contact'
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "404":
          description: Contact not found
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Update Contact
      tags:
      - contact
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"